Seller details and eager loading

+++ SalesWebMvc/Services/SellerService.cs
@@ -2,6 +2,7 @@
 using SalesWebMvc.Data;
 using NuGet.Protocol.Plugins;
 using SalesWebMvc.Models.ViewModels;
+using Microsoft.EntityFrameworkCore;
 
 namespace SalesWebMvc.Services
 {
@@ -21,21 +22,24 @@ namespace SalesWebMvc.Services
 
         public void Insert(Seller obj) 
         {
-            obj.Department = _context.Department.FirstOrDefault(d => d.Id == obj.DepartmentId);
+            obj.Department = _context.Department.FirstOrDefault(d => d.Id == obj.DepartmentId) ?? new Department();
             _context.Add(obj);
             _context.SaveChanges();
         }
 
         public Seller FindById(int id)
         {
-            return _context.Seller.FirstOrDefault(d => d.Id == id);
+            return _context.Seller.Include(obj => obj.Department).FirstOrDefault(d => d.Id == id) ?? new Seller();
         }
         
         public void Remove(int id) 
         {
             var obj = _context.Seller.Find(id);
-            _context.Seller.Remove(obj);
-            _context.SaveChanges();
+            if (obj != null)
+            {
+                _context.Seller.Remove(obj);
+                _context.SaveChanges();
+            }
         }
     }
 }
